Migrate more users endpoints
