+++ server/service/handler_test.go
@@ -43,14 +43,6 @@ func TestAPIRoutes(t *testing.T) {
 			uri:  "/api/v1/fleet/users",
 		},
 		{
-			verb: "GET",
-			uri:  "/api/v1/fleet/users/1",
-		},
-		{
-			verb: "PATCH",
-			uri:  "/api/v1/fleet/users/1",
-		},
-		{
 			verb: "POST",
 			uri:  "/api/v1/fleet/login",
 		},
