+++ server/service/service_users.go
@@ -7,7 +7,6 @@ import (
 	"time"
 
 	"github.com/fleetdm/fleet/v4/server"
-	"github.com/fleetdm/fleet/v4/server/authz"
 
 	"github.com/fleetdm/fleet/v4/server/contexts/ctxerr"
 	"github.com/fleetdm/fleet/v4/server/contexts/viewer"
@@ -85,157 +84,6 @@ func (svc *Service) newUser(ctx context.Context, p fleet.UserPayload) (*fleet.Us
 	return user, nil
 }
 
-func (svc *Service) ModifyUser(ctx context.Context, userID uint, p fleet.UserPayload) (*fleet.User, error) {
-	if err := svc.authz.Authorize(ctx, &fleet.User{}, fleet.ActionRead); err != nil {
-		return nil, err
-	}
-
-	user, err := svc.User(ctx, userID)
-	if err != nil {
-		return nil, err
-	}
-
-	if err := svc.authz.Authorize(ctx, user, fleet.ActionWrite); err != nil {
-		return nil, err
-	}
-
-	if p.GlobalRole != nil || p.Teams != nil {
-		if err := svc.authz.Authorize(ctx, user, fleet.ActionWriteRole); err != nil {
-			return nil, err
-		}
-	}
-	if p.Name != nil {
-		user.Name = *p.Name
-	}
-
-	if p.Email != nil && *p.Email != user.Email {
-		err = svc.modifyEmailAddress(ctx, user, *p.Email, p.Password)
-		if err != nil {
-			return nil, err
-		}
-	}
-
-	if p.Position != nil {
-		user.Position = *p.Position
-	}
-
-	if p.GravatarURL != nil {
-		user.GravatarURL = *p.GravatarURL
-	}
-
-	if p.SSOEnabled != nil {
-		user.SSOEnabled = *p.SSOEnabled
-	}
-
-	currentUser := authz.UserFromContext(ctx)
-
-	if p.GlobalRole != nil && *p.GlobalRole != "" {
-		if currentUser.GlobalRole == nil {
-			return nil, ctxerr.New(ctx, "Cannot edit global role as a team member")
-		}
-
-		if p.Teams != nil && len(*p.Teams) > 0 {
-			return nil, fleet.NewInvalidArgumentError("teams", "may not be specified with global_role")
-		}
-		user.GlobalRole = p.GlobalRole
-		user.Teams = []fleet.UserTeam{}
-	} else if p.Teams != nil {
-		if !isAdminOfTheModifiedTeams(currentUser, user.Teams, *p.Teams) {
-			return nil, ctxerr.New(ctx, "Cannot modify teams in that way")
-		}
-		user.Teams = *p.Teams
-		user.GlobalRole = nil
-	}
-
-	err = svc.saveUser(ctx, user)
-	if err != nil {
-		return nil, err
-	}
-
-	return user, nil
-}
-
-func isAdminOfTheModifiedTeams(currentUser *fleet.User, originalUserTeams, newUserTeams []fleet.UserTeam) bool {
-	// If the user is of the right global role, then they can modify the teams
-	if currentUser.GlobalRole != nil && (*currentUser.GlobalRole == fleet.RoleAdmin || *currentUser.GlobalRole == fleet.RoleMaintainer) {
-		return true
-	}
-
-	// otherwise, gather the resulting teams
-	resultingTeams := make(map[uint]string)
-	for _, team := range newUserTeams {
-		resultingTeams[team.ID] = team.Role
-	}
-
-	// and see which ones were removed or changed from the original
-	teamsAffected := make(map[uint]struct{})
-	for _, team := range originalUserTeams {
-		if resultingTeams[team.ID] != team.Role {
-			teamsAffected[team.ID] = struct{}{}
-		}
-	}
-
-	// then gather the teams the current user is admin for
-	currentUserTeamAdmin := make(map[uint]struct{})
-	for _, team := range currentUser.Teams {
-		if team.Role == fleet.RoleAdmin {
-			currentUserTeamAdmin[team.ID] = struct{}{}
-		}
-	}
-
-	// and let's check that the teams that were either removed or changed are also teams this user is an admin of
-	for teamID := range teamsAffected {
-		if _, ok := currentUserTeamAdmin[teamID]; !ok {
-			return false
-		}
-	}
-
-	return true
-}
-
-func (svc *Service) modifyEmailAddress(ctx context.Context, user *fleet.User, email string, password *string) error {
-	// password requirement handled in validation middleware
-	if password != nil {
-		err := user.ValidatePassword(*password)
-		if err != nil {
-			return fleet.NewPermissionError("incorrect password")
-		}
-	}
-	random, err := server.GenerateRandomText(svc.config.App.TokenKeySize)
-	if err != nil {
-		return err
-	}
-	token := base64.URLEncoding.EncodeToString([]byte(random))
-	err = svc.ds.PendingEmailChange(ctx, user.ID, email, token)
-	if err != nil {
-		return err
-	}
-	config, err := svc.AppConfig(ctx)
-	if err != nil {
-		return err
-	}
-
-	changeEmail := fleet.Email{
-		Subject: "Confirm Fleet Email Change",
-		To:      []string{email},
-		Config:  config,
-		Mailer: &mail.ChangeEmailMailer{
-			Token:    token,
-			BaseURL:  template.URL(config.ServerSettings.ServerURL + svc.config.Server.URLPrefix),
-			AssetURL: getAssetURL(),
-		},
-	}
-	return svc.mailService.SendEmail(changeEmail)
-}
-
-func (svc *Service) DeleteUser(ctx context.Context, id uint) error {
-	if err := svc.authz.Authorize(ctx, &fleet.User{ID: id}, fleet.ActionWrite); err != nil {
-		return err
-	}
-
-	return svc.ds.DeleteUser(ctx, id)
-}
-
 func (svc *Service) ChangeUserEmail(ctx context.Context, token string) (string, error) {
 	vc, ok := viewer.FromContext(ctx)
 	if !ok {
@@ -247,14 +95,6 @@ func (svc *Service) ChangeUserEmail(ctx context.Context, token string) (string, 
 	}
 
 	return svc.ds.ConfirmPendingEmailChange(ctx, vc.UserID(), token)
-}
-
-func (svc *Service) User(ctx context.Context, id uint) (*fleet.User, error) {
-	if err := svc.authz.Authorize(ctx, &fleet.User{ID: id}, fleet.ActionRead); err != nil {
-		return nil, err
-	}
-
-	return svc.ds.UserByID(ctx, id)
 }
 
 func (svc *Service) UserUnauthorized(ctx context.Context, id uint) (*fleet.User, error) {
@@ -402,34 +242,6 @@ func (svc *Service) PerformRequiredPasswordReset(ctx context.Context, password s
 	return user, nil
 }
 
-func (svc *Service) RequirePasswordReset(ctx context.Context, uid uint, require bool) (*fleet.User, error) {
-	if err := svc.authz.Authorize(ctx, &fleet.User{ID: uid}, fleet.ActionWrite); err != nil {
-		return nil, err
-	}
-
-	user, err := svc.ds.UserByID(ctx, uid)
-	if err != nil {
-		return nil, ctxerr.Wrap(ctx, err, "loading user by ID")
-	}
-	if user.SSOEnabled {
-		return nil, ctxerr.New(ctx, "password reset for single sign on user not allowed")
-	}
-	// Require reset on next login
-	user.AdminForcedPasswordReset = require
-	if err := svc.saveUser(ctx, user); err != nil {
-		return nil, ctxerr.Wrap(ctx, err, "saving user")
-	}
-
-	if require {
-		// Clear all of the existing sessions
-		if err := svc.DeleteSessionsForUser(ctx, user.ID); err != nil {
-			return nil, ctxerr.Wrap(ctx, err, "deleting user sessions")
-		}
-	}
-
-	return user, nil
-}
-
 func (svc *Service) RequestPasswordReset(ctx context.Context, email string) error {
 	// skipauth: No viewer context available. The user is locked out of their
 	// account and trying to reset their password.
@@ -482,11 +294,4 @@ func (svc *Service) RequestPasswordReset(ctx context.Context, email string) erro
 	}
 
 	return svc.mailService.SendEmail(resetEmail)
-}
-
-// saves user in datastore.
-// doesn't need to be exposed to the transport
-// the service should expose actions for modifying a user instead
-func (svc *Service) saveUser(ctx context.Context, user *fleet.User) error {
-	return svc.ds.SaveUser(ctx, user)
 }
