+++ server/service/transport_sessions.go
@@ -18,28 +18,12 @@ func decodeGetInfoAboutSessionRequest(ctx context.Context, r *http.Request) (int
 	return getInfoAboutSessionRequest{ID: uint(id)}, nil
 }
 
-func decodeGetInfoAboutSessionsForUserRequest(ctx context.Context, r *http.Request) (interface{}, error) {
-	id, err := uintFromRequest(r, "id")
-	if err != nil {
-		return nil, err
-	}
-	return getInfoAboutSessionsForUserRequest{ID: uint(id)}, nil
-}
-
 func decodeDeleteSessionRequest(ctx context.Context, r *http.Request) (interface{}, error) {
 	id, err := uintFromRequest(r, "id")
 	if err != nil {
 		return nil, err
 	}
 	return deleteSessionRequest{ID: uint(id)}, nil
-}
-
-func decodeDeleteSessionsForUserRequest(ctx context.Context, r *http.Request) (interface{}, error) {
-	id, err := uintFromRequest(r, "id")
-	if err != nil {
-		return nil, err
-	}
-	return deleteSessionsForUserRequest{ID: uint(id)}, nil
 }
 
 func decodeLoginRequest(ctx context.Context, r *http.Request) (interface{}, error) {
