+++ server/service/transport_sessions_test.go
@@ -27,22 +27,6 @@ func TestDecodeGetInfoAboutSessionRequest(t *testing.T) {
 	)
 }
 
-func TestDecodeGetInfoAboutSessionsForUserRequest(t *testing.T) {
-	router := mux.NewRouter()
-	router.HandleFunc("/api/v1/fleet/user/{id}/sessions", func(writer http.ResponseWriter, request *http.Request) {
-		r, err := decodeGetInfoAboutSessionsForUserRequest(context.Background(), request)
-		assert.Nil(t, err)
-
-		params := r.(getInfoAboutSessionsForUserRequest)
-		assert.Equal(t, uint(1), params.ID)
-	}).Methods("GET")
-
-	router.ServeHTTP(
-		httptest.NewRecorder(),
-		httptest.NewRequest("GET", "/api/v1/fleet/users/1/sessions", nil),
-	)
-}
-
 func TestDecodeDeleteSessionRequest(t *testing.T) {
 	router := mux.NewRouter()
 	router.HandleFunc("/api/v1/fleet/sessions/{id}", func(writer http.ResponseWriter, request *http.Request) {
@@ -56,22 +40,6 @@ func TestDecodeDeleteSessionRequest(t *testing.T) {
 	router.ServeHTTP(
 		httptest.NewRecorder(),
 		httptest.NewRequest("DELETE", "/api/v1/fleet/sessions/1", nil),
-	)
-}
-
-func TestDecodeDeleteSessionsForUserRequest(t *testing.T) {
-	router := mux.NewRouter()
-	router.HandleFunc("/api/v1/fleet/user/{id}/sessions", func(writer http.ResponseWriter, request *http.Request) {
-		r, err := decodeDeleteSessionsForUserRequest(context.Background(), request)
-		assert.Nil(t, err)
-
-		params := r.(deleteSessionsForUserRequest)
-		assert.Equal(t, uint(1), params.ID)
-	}).Methods("DELETE")
-
-	router.ServeHTTP(
-		httptest.NewRecorder(),
-		httptest.NewRequest("DELETE", "/api/v1/fleet/users/1/sessions", nil),
 	)
 }
